Got filtering working
Former-commit-id: 96981dddf3c347fc6f94846d3e0014b3c2413509

+++ engine/Source/Holodeck/General/Private/Octree.cpp
@@ -33,7 +33,7 @@ void Octree::makeOctree(FVector center, float size, UWorld* World, TArray<Octree
     bool occup = World->SweepSingleByChannel(hit, center, center+offset, FQuat::Identity, ECollisionChannel::ECC_WorldStatic, FCollisionShape::MakeBox(FVector(size/2)), params);
     occup = hit.bStartPenetrating;
 
-    static int cornerSize = .1;
+    static float cornerSize = .1;
 
     // if it's occupied
 	if(occup){
